split simtest into multiple build target
