+++ crates/walrus-simtest/src/lib.rs
@@ -1,0 +1,10 @@
+// Copyright (c) Walrus Foundation
+// SPDX-License-Identifier: Apache-2.0
+
+#![recursion_limit = "256"]
+
+//! Contains test utilities for the Walrus simulation tests.
+//! This crate provides utilities for testing Walrus in a simulated environment.
+
+/// Contains test utilities for the Walrus simulation tests.
+pub mod test_utils;
